Move remaining krate endpoints to krate::metadata::*
