+++ src/lib.rs
@@ -157,9 +157,12 @@ pub fn middleware(app: Arc<App>) -> MiddlewareBuilder {
     api_router.get("/versions/:version_id", C(version::show));
 
     // Routes used by the frontend
-    api_router.get("/crates/:crate_id", C(krate::show));
+    api_router.get("/crates/:crate_id", C(krate::metadata::show));
     api_router.get("/crates/:crate_id/:version", C(version::show));
-    api_router.get("/crates/:crate_id/:version/readme", C(krate::readme));
+    api_router.get(
+        "/crates/:crate_id/:version/readme",
+        C(krate::metadata::readme),
+    );
     api_router.get(
         "/crates/:crate_id/:version/dependencies",
         C(version::dependencies),
@@ -173,7 +176,7 @@ pub fn middleware(app: Arc<App>) -> MiddlewareBuilder {
         "/crates/:crate_id/downloads",
         C(krate::downloads::downloads),
     );
-    api_router.get("/crates/:crate_id/versions", C(krate::versions));
+    api_router.get("/crates/:crate_id/versions", C(krate::metadata::versions));
     api_router.put("/crates/:crate_id/follow", C(krate::follow::follow));
     api_router.delete("/crates/:crate_id/follow", C(krate::follow::unfollow));
     api_router.get("/crates/:crate_id/following", C(krate::follow::following));
@@ -181,7 +184,7 @@ pub fn middleware(app: Arc<App>) -> MiddlewareBuilder {
     api_router.get("/crates/:crate_id/owner_user", C(krate::owners::owner_user));
     api_router.get(
         "/crates/:crate_id/reverse_dependencies",
-        C(krate::reverse_dependencies),
+        C(krate::metadata::reverse_dependencies),
     );
     api_router.get("/keywords", C(keyword::index));
     api_router.get("/keywords/:keyword_id", C(keyword::show));
@@ -205,7 +208,7 @@ pub fn middleware(app: Arc<App>) -> MiddlewareBuilder {
         "/me/crate_owner_invitations/:crate_id",
         C(crate_owner_invitation::handle_invite),
     );
-    api_router.get("/summary", C(krate::summary));
+    api_router.get("/summary", C(krate::metadata::summary));
     api_router.put("/confirm/:email_token", C(user::confirm_user_email));
     api_router.put("/users/:user_id/resend", C(user::regenerate_token_and_send));
     let api_router = Arc::new(R404(api_router));
